// src/types/apps/chatTypes.ts
export type UserChatType = {
  _id: string
  firstName: string
  username: string
  lastName: string
  gender: string
  avatar: string | null
  role: string
  isActive: boolean
}

export type LastMessageType = {
  _id: string
  user: string
  content: string
  reply_to: string
  isDelete: boolean
  images: string[]
  createdAt: string
  updatedAt: string
}

export type ChatType = {
  _id?: string
  createdAt?: string
  updatedAt?: string
  user: UserChatType
  lastMessage?: LastMessageType
  unReadMessages: number
}
// ...
export type SendMsgComponentType = {
  selectedChat: ChatType
  setMessages: React.Dispatch<React.SetStateAction<MessageType[]>>
  setSelectedChat: React.Dispatch<React.SetStateAction<null | ChatType>>,
  replyTo: MessageType | null
  setReplyTo: React.Dispatch<React.SetStateAction<MessageType | null>>
}

export type ChatLogType = {
  hidden: boolean
  selectedChat: ChatType
  page: number
  messages: MessageType[]
  hasMore: boolean
  getChatContent: (id?: string, page?: number | null) => void
  replyTo: MessageType | null
  setReplyTo: React.Dispatch<React.SetStateAction<MessageType | null>>
}

export type MessageType = {
  _id: string
  user: UserChatType
  content: string
  reply_to: string
  isDelete: false
  images: string[]
  reacts: { [id: string]: string }
  seen: { [id: string]: string }
  createdAt: string
  updatedAt: string
}